+++ apps/boruta_identity/assets/wallet/src/components/KeySelect.vue
@@ -8,34 +8,39 @@
         @consent="removeKeyConsent"
       />
       <h2>Select a key</h2>
-      <div class="ui error message" v-if="error">{{ error }}</div>
-      <div class="ui cards" v-else>
+      <div class="ui error message" v-if="keyPromptError">{{ keyPromptError }}</div>
+      <div class="ui cards">
         <div class="card" v-for="[identifier, did] in keys">
           <div class="content">
             <div class="header">
               {{ identifier }}
             </div>
-            <div class="description" v-if="selectedKeys.includes(identifier)">
-              <p class="ui warning message"><em>key confirmed</em></p>
-            </div>
-          </div>
-          <div class="extra content">
-            <div class="ui two buttons">
-              <button class="ui basic blue button" @click="$emit('selected', identifier, did)">Use this key</button>
-              <button class="ui basic red button" @click="deleteKey(identifier)">Delete</button>
-            </div>
-          </div>
-        </div>
-        <div class="card" v-if="!requestedKey">
-          <div class="content">
-            <div class="header">
+            <div class="description">
+              <p class="ui warning message" v-if="selectedKeys.includes(identifier)"><em>key confirmed</em></p>
               <div class="ui form">
-                <input type="text" v-model="newIdentifier" placeholder="Key identifier"/>
+                <input type="hidden" :value="identifier" />
+                <input type="password" v-model="passwords[identifier]" placeholder="Key password" autocomplete="current-password" @keyup.enter="selectKey(identifier)" />
               </div>
             </div>
           </div>
           <div class="extra content">
-            <button :disabled="!newIdentifier" class="ui fluid basic blue button" @click="$emit('selected', newIdentifier)">Add a new key</button>
+            <div class="ui two buttons">
+              <button :disabled="!passwords[identifier]" class="ui basic blue button" @click="selectKey(identifier)">Use this key</button>
+              <button class="ui basic red button" @click="deleteKey(identifier)">Delete</button>
+            </div>
+          </div>
+        </div>
+        <div class="card">
+          <div class="content">
+            <div class="header">
+              <div class="ui form">
+                <input type="text" v-model="newIdentifier" placeholder="Key identifier"/>
+                <input type="password" v-model="newPassword" placeholder="Key password" autocomplete="new-password"/>
+              </div>
+            </div>
+          </div>
+          <div class="extra content">
+            <button :disabled="!newIdentifier || !newPassword" class="ui fluid basic blue button" @click="$emit('selected', newIdentifier, null, newPassword)">Add a new key</button>
           </div>
         </div>
       </div>
@@ -47,7 +52,8 @@
 
 <script>
 import { defineComponent } from 'vue'
-import { BorutaOauth, KeyStore, CustomEventHandler } from 'boruta-client'
+import { KeyStore, CustomEventHandler } from 'boruta-client'
+import { EbsiWallet } from '@cef-ebsi/wallet-lib'
 import { storage } from '../store'
 
 import Consent from './Consent.vue'
@@ -63,10 +69,13 @@ export default defineComponent({
     return {
       keys: [],
       newIdentifier: null,
+      newPassword: null,
       removeKeyConsentEventKey: null,
       requestedKey: null,
+      requestedDid: null,
       selectedKeys: [],
-      error: null,
+      keyPromptError: null,
+      passwords: {},
       keyStore
     }
   },
@@ -80,29 +89,13 @@ export default defineComponent({
       })
     })
 
-    await Promise.all(this.keys.map(async ([identifier]) => {
-      return [identifier, await this.keyStore.extractDid(identifier)]
-    })).then(keys => {
-      this.keys = keys
-      const key = keys.find(([identifier, did]) => {
-        console.log(did)
-        try {
-          return this.$route.query.client_id == did ||
-            JSON.parse(this.$route.query.credential_offer).client_id == did
-        } catch (_error) {
-          return false
-        }
-      })
-
-      if (key) {
-        this.requestedKey = key[0]
-      }
-    })
+    this.requestedDid = this.requestedDidFromRoute()
+    this.requestedKey = this.requestedDid
 
     if (this.requestedKey) {
       const keySelections = localStorage.getItem('keySelection')
       if (keySelections) {
-        keySelections.split('|').forEach(keySelection => {
+        keySelections.split('|').filter(Boolean).forEach(keySelection => {
           const keySelectedAt = keySelection.split('~')[0]
           const selectedKey = keySelection.split('~')[1]
           if (parseInt(keySelectedAt) + 60000 > Date.now()) {
@@ -111,13 +104,36 @@ export default defineComponent({
             localStorage.setItem('keySelection', keySelections.replace('|' + keySelection, ''))
           }
         })
-      } else {
-        this.keys = []
-        this.error = 'Cannot confirm requested key.'
       }
     }
   },
   methods: {
+    async selectKey (identifier) {
+      const password = this.passwords[identifier]
+
+      if (!password) return
+
+      this.keyPromptError = null
+
+      try {
+        const keyPair = await this.keyStore.keyPair(identifier, password)
+        if (!keyPair?.publicKeyJwk) {
+          this.keyPromptError = 'Could not unlock selected key.'
+          return
+        }
+
+        const did = EbsiWallet.createDid('NATURAL_PERSON', keyPair.publicKeyJwk)
+
+        if (this.requestedDid && did != this.requestedDid) {
+          this.keyPromptError = 'Selected key does not match the requested key.'
+          return
+        }
+
+        this.$emit('selected', identifier, did, password)
+      } catch (_error) {
+        this.keyPromptError = 'Could not unlock selected key.'
+      }
+    },
     deleteKey (identifier) {
       this.keyStore.removeKey(identifier).then(keys => {
         this.keys = keys.map(key => [key])
@@ -129,6 +145,22 @@ export default defineComponent({
     },
     abortKeyConsent () {
       this.removeKeyConsentEventKey = null
+    },
+    requestedDidFromRoute () {
+      if (this.$route.query.client_id) {
+        return this.didOrNull(this.$route.query.client_id)
+      }
+
+      try {
+        return this.didOrNull(JSON.parse(this.$route.query.credential_offer).client_id)
+      } catch (_error) {
+        return null
+      }
+    },
+    didOrNull (value) {
+      if (typeof value != 'string') return null
+
+      return value.startsWith('did:') ? value : null
     }
   }
 })